Handle absolute CMAKE_INSTALL_PREFIX
This allows CMAKE_INSTALL_PREFIX to be both relative and absolute when using source-built AGX applications. This is required if building out-of-source, since a relative CMAKE_INSTALL_PREFIX is interpreted as being relative to the source directory, while AGX_DIR in the code above is relative to the binary directory.